+++ plugins/unplugin/tsdown.config.ts
@@ -5,13 +5,15 @@ export default defineConfig(
     {
         entry: {
             // public
+            "webpack/index": "./src/bundlers/webpack/index.ts",
+            "rspack/index": "./src/bundlers/rspack/index.ts",
             "rollup/index": "./src/bundlers/rollup/index.ts",
             "vite/index": "./src/bundlers/vite/index.ts",
-            "webpack/index": "./src/bundlers/webpack/index.ts",
             // internal
+            "webpack/create": "./src/bundlers/webpack/create.ts",
+            "rspack/create": "./src/bundlers/rspack/create.ts",
             "rollup/create": "./src/bundlers/rollup/create.ts",
             "vite/create": "./src/bundlers/vite/create.ts",
-            "webpack/create": "./src/bundlers/webpack/create.ts",
             // required
             "plugins/vite/hmr/hmr": "./src/plugins/vite/hmr/hmr.ts",
             // extra